Second Commit
Working dynamics with no hit detection implemented

+++ src/interfaces/MovableObject.java
@@ -2,5 +2,5 @@ package interfaces;
 
 public interface MovableObject extends PhysicalObject {
 
-	public void update(int time);
+	public void update(double time);
 }
